Add error messages for unsupported scenarios.
Improve testing.
Split classes into separate files.
