+++ src/Microsoft.EntityFrameworkCore.Design/Design/OperationExecutor.cs
@@ -17,7 +17,7 @@ namespace Microsoft.EntityFrameworkCore.Design
     /// <summary>
     ///     A version-resilient, AppDomain-and-reflection-friendly facade for command operations.
     /// </summary>
-    public partial class OperationExecutor
+    public class OperationExecutor : DesignMarshalByRefObject
     {
         private readonly LazyRef<DbContextOperations> _contextOperations;
         private readonly LazyRef<DatabaseOperations> _databaseOperations;
@@ -399,7 +399,7 @@ namespace Microsoft.EntityFrameworkCore.Design
         private void DropDatabaseImpl(string contextType)
             => _contextOperations.Value.DropDatabase(contextType);
 
-        public abstract partial class OperationBase
+        public abstract class OperationBase : DesignMarshalByRefObject
         {
             private readonly IOperationResultHandler _resultHandler;
 
@@ -439,14 +439,5 @@ namespace Microsoft.EntityFrameworkCore.Design
             }
         }
     }
-
-#if NET451
-    partial class OperationExecutor : MarshalByRefObject
-    {
-        partial class OperationBase : MarshalByRefObject
-        {
-        }
-    }
-#endif
 }
