+++ src/Microsoft.EntityFrameworkCore.Design/Design/OperationResultHandler.cs
@@ -1,9 +1,9 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using Microsoft.EntityFrameworkCore.Design;
+using Microsoft.EntityFrameworkCore.Design.Internal;
 
-namespace Microsoft.EntityFrameworkCore.Tools.Internal
+namespace Microsoft.EntityFrameworkCore.Design
 {
     public class OperationResultHandler : DesignMarshalByRefObject, IOperationResultHandler
     {
