+++ src/Tools.Console/Internal/AppDomainOperationExecutor.net451.cs
