+++ src/Tools.Console/Internal/IOperationResultHandler.net451.cs
@@ -2,6 +2,7 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 #if NET451
+// ReSharper disable once CheckNamespace
 namespace Microsoft.EntityFrameworkCore.Design
 {
     public interface IOperationResultHandler
