+++ src/Tools.Console/Internal/ReflectionOperationExecutor.netcore.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+#if NETCOREAPP1_0
 using System;
 using System.Collections;
 using System.Collections.Generic;
@@ -55,3 +56,4 @@ namespace Microsoft.EntityFrameworkCore.Tools.Internal
         }
     }
 }
+#endif
