+++ src/Tools.Console/Tools.Console.csproj
@@ -8,7 +8,7 @@
     <ProjectGuid>{74A7774B-270E-4D7F-8735-BECB05CACA71}</ProjectGuid>
     <OutputType>Exe</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Microsoft.EntityFrameworkCore.Tools</RootNamespace>
+    <RootNamespace>Microsoft.EntityFrameworkCore</RootNamespace>
     <AssemblyName>ef</AssemblyName>
     <TargetFrameworkVersion>v4.5.1</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
@@ -23,6 +23,8 @@
     <DefineConstants>TRACE;DEBUG;NET451</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
+    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
+    <NoWarn>1591;</NoWarn>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
     <PlatformTarget>AnyCPU</PlatformTarget>
@@ -36,47 +38,60 @@
   <ItemGroup>
     <Reference Include="System" />
     <Reference Include="System.Core" />
-    <Reference Include="System.Xml.Linq" />
-    <Reference Include="System.Data.DataSetExtensions" />
     <Reference Include="Microsoft.CSharp" />
-    <Reference Include="System.Data" />
-    <Reference Include="System.Net.Http" />
-    <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Content Include="..\..\tools\Resources.tt">
+      <Link>Properties\Resources.tt</Link>
+      <Generator>TextTemplatingFileGenerator</Generator>
+      <LastGenOutput>Resources.cs</LastGenOutput>
+      <CustomToolNamespace>Microsoft.EntityFrameworkCore.Internal</CustomToolNamespace>
+    </Content>
+    <EmbeddedResource Include="Properties\ToolsStrings.resx">
+      <LogicalName>Tools.Console.Properties.ToolsStrings.resources</LogicalName>
+    </EmbeddedResource>
+    <Compile Include="..\..\tools\Resources.cs">
+      <Link>Properties\Resources.cs</Link>
+      <AutoGen>True</AutoGen>
+      <DesignTime>True</DesignTime>
+      <DependentUpon>Resources.tt</DependentUpon>
+    </Compile>
     <Compile Include="..\Shared\CodeAnnotations.cs">
       <Link>CodeAnnotations.cs</Link>
     </Compile>
     <Compile Include="..\Tools.Shared\CommandLineApplicationExtensions.cs">
       <Link>Extensions\CommandLineApplicationExtensions.cs</Link>
     </Compile>
-    <Compile Include="Commands\EfCommand.cs" />
-    <Compile Include="Commands\MigrationsCommand.cs" />
-    <Compile Include="Commands\DbContextCommand.cs" />
-    <Compile Include="Commands\DatabaseCommand.cs" />
-    <Compile Include="Commands\ICommand.cs" />
-    <Compile Include="Commands\MigrationsScriptCommand.cs" />
-    <Compile Include="Internal\IOperationLogHandler.cs" />
-    <Compile Include="Internal\IOperationResultHandler.cs" />
-    <Compile Include="Internal\IOperationExecutor.cs" />
-    <Compile Include="Internal\OperationLogHandler.cs" />
-    <Compile Include="Internal\OperationResultHandler.cs" />
-    <Compile Include="Internal\ReflectionOperationExecutor.cs" />
-    <Compile Include="Internal\DesignMarshalByRefObject.cs" />
-    <Compile Include="Internal\OperationExecutorBase.cs" />
-    <Compile Include="OperationErrorException.cs" />
-    <Compile Include="Reporter.cs" />
-    <Compile Include="Commands\DatabaseDropCommand.cs" />
-    <Compile Include="Commands\DatabaseUpdateCommand.cs" />
-    <Compile Include="Commands\DbContextListCommand.cs" />
-    <Compile Include="Commands\DbContextScaffoldCommand.cs" />
-    <Compile Include="CommandLineOptions.cs" />
+    <Compile Include="Tools\Commands\EfCommand.cs" />
+    <Compile Include="Tools\Commands\MigrationsCommand.cs" />
+    <Compile Include="Tools\Commands\DbContextCommand.cs" />
+    <Compile Include="Tools\Commands\DatabaseCommand.cs" />
+    <Compile Include="Tools\Commands\ICommand.cs" />
+    <Compile Include="Tools\Commands\MigrationsScriptCommand.cs" />
+    <Compile Include="Design\IOperationLogHandler.net451.cs" />
+    <Compile Include="Design\IOperationResultHandler.net451.cs" />
+    <Compile Include="Design\Internal\IOperationExecutor.cs" />
+    <Compile Include="Design\OperationLogHandler.cs" />
+    <Compile Include="Design\OperationResultHandler.cs" />
+    <Compile Include="Design\Internal\ReflectionOperationExecutor.cs" />
+    <Compile Include="Design\Internal\DesignMarshalByRefObject.cs" />
+    <Compile Include="Design\Internal\OperationExecutorBase.cs" />
+    <Compile Include="Tools\OperationErrorException.cs" />
+    <Compile Include="Properties\ToolsStrings.Designer.cs">
+      <DependentUpon>ToolsStrings.resx</DependentUpon>
+    </Compile>
+    <Compile Include="Tools\Reporter.cs" />
+    <Compile Include="Tools\Commands\DatabaseDropCommand.cs" />
+    <Compile Include="Tools\Commands\DatabaseUpdateCommand.cs" />
+    <Compile Include="Tools\Commands\DbContextListCommand.cs" />
+    <Compile Include="Tools\Commands\DbContextScaffoldCommand.cs" />
+    <Compile Include="Tools\CommandLineOptions.cs" />
     <Compile Include="Extensions\AnsiColorExtensions.cs" />
-    <Compile Include="Internal\AppDomainOperationExecutor.cs" />
-    <Compile Include="Commands\MigrationsAddCommand.cs" />
-    <Compile Include="Commands\MigrationsListCommand.cs" />
-    <Compile Include="Commands\MigrationsRemoveCommand.cs" />
-    <Compile Include="Program.cs" />
+    <Compile Include="Design\Internal\AppDomainOperationExecutor.net451.cs" />
+    <Compile Include="Tools\Commands\MigrationsAddCommand.cs" />
+    <Compile Include="Tools\Commands\MigrationsListCommand.cs" />
+    <Compile Include="Tools\Commands\MigrationsRemoveCommand.cs" />
+    <Compile Include="Tools\Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
@@ -88,7 +103,9 @@
       <Name>Microsoft.EntityFrameworkCore.Design.Abstractions</Name>
     </ProjectReference>
   </ItemGroup>
-  <ItemGroup />
+  <ItemGroup>
+    <Service Include="{508349B6-6B84-4DF5-91F0-309BEEBAD82D}" />
+  </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
