+++ test/Tools.Console.FunctionalTests/Tools.Console.FunctionalTests.xproj
@@ -6,13 +6,16 @@
   </PropertyGroup>
   <Import Project="$(VSToolsPath)\DotNet\Microsoft.DotNet.Props" Condition="'$(VSToolsPath)' != ''" />
   <PropertyGroup Label="Globals">
-    <ProjectGuid>4682c48a-94a2-42df-afbc-1c3a6cd7195a</ProjectGuid>
-    <RootNamespace>Microsoft.EntityFrameworkCore.Design</RootNamespace>
+    <ProjectGuid>667be700-2b7d-4e6d-bffb-1bffef51d078</ProjectGuid>
+    <RootNamespace>Microsoft.EntityFrameworkCore.Tools.FunctionalTests</RootNamespace>
     <BaseIntermediateOutputPath Condition="'$(BaseIntermediateOutputPath)'=='' ">.\obj</BaseIntermediateOutputPath>
     <OutputPath Condition="'$(OutputPath)'=='' ">.\bin\</OutputPath>
   </PropertyGroup>
   <PropertyGroup>
     <SchemaVersion>2.0</SchemaVersion>
   </PropertyGroup>
+  <ItemGroup>
+    <Service Include="{82a7f48d-3b50-4b1e-b82e-3ada8210c358}" />
+  </ItemGroup>
   <Import Project="$(VSToolsPath)\DotNet\Microsoft.DotNet.targets" Condition="'$(VSToolsPath)' != ''" />
 </Project>
