+++ test/Tools.Console.FunctionalTests/project.json
@@ -15,6 +15,14 @@
   },
   "testRunner": "xunit",
   "frameworks": {
-    "net451": {}
+    "net451": { },
+    "netcoreapp1.0": {
+      "dependencies": {
+        "Microsoft.NETCore.App": {
+          "version": "1.0.0-*",
+          "type": "platform"
+        }
+      }
+    }
   }
 }
