+++ test/Tools.DotNet.FunctionalTests/TestProjects/NetStandardClassLibrary/project.json.ignore
@@ -4,6 +4,9 @@
     "Microsoft.EntityFrameworkCore.Sqlite": {
       "target": "project"
     },
+    "Microsoft.EntityFrameworkCore.Design": {
+      "target": "project"
+    },
     "NETStandard.Library": "1.6.0-*"
   },
   "frameworks": {
